Cleaned up files.
