+++ client/src/components/AccountForm.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 import { signUpOrIn } from '../lib/accountApi';
-import './AccountForm.css';
+import './componentsCSS/AccountForm.css';
 
 export default function AccountForm({ action, onSignIn }) {
   const navigate = useNavigate();
