+++ client/src/components/Carousel.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, useCallback } from 'react';
 import { AiOutlineLeft, AiOutlineRight } from 'react-icons/ai';
 import { RiGamepadFill, RiGamepadLine } from 'react-icons/ri';
-import './Carousel.css';
+import './componentsCSS/Carousel.css';
 import { Link } from 'react-router-dom';
 
 export default function Carousel() {
