+++ client/src/components/componentsCSS/SearchBar.css
