+++ client/src/pages/Cart.jsx
@@ -1,11 +1,11 @@
-import './Cart.css';
+import './pagesCSS/Cart.css';
 import { useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
-import { GiTabletopPlayers } from 'react-icons/gi'
 import { removeFromCart } from '../lib/cartApi';
 import totalPrice from '../lib/checkout';
 import { totalQuantity } from '../lib/checkout';
 import { getCart } from '../lib/cartApi';
+import CartProducts from '../components/CartProducts';
 
 export default function Cart() {
   const [products, setProducts] = useState([]);
@@ -73,41 +73,6 @@ export default function Cart() {
             <div className="d-flex flex-column justify-content-center align-items-center">
               <button className="checkout-button btn btn-danger mb-3" onClick={emptyCart}>Empty Cart</button>
               <button className="checkout-button btn btn-success">Checkout</button>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function CartProducts({ product, update }) {
-  const { name, price, description, minPlayers, maxPlayers, imageUrl, quantity, cartId } = product;
-
-  function handleRemoveProduct(event) {
-    event.preventDefault();
-    const removeProduct = {
-      cartId
-    };
-    removeFromCart(removeProduct);
-    update(cartId)
-  };
-
-  return (
-    <div className="container">
-      <div className="card shadow-sm">
-        <div className="card-body">
-          <div className="row mb-4">
-            <div className="col-12 col-md-5 mb-3 mb-md-0">
-              <img src={imageUrl} alt={name} className="cartImage img-fluid mx-auto" />
-            </div>
-            <div className="col-12 col-md-7">
-              <h2>{name}</h2>
-              <h5 className="text-secondary">${price}</h5>
-              <p className="description card-text text-truncate">{description}</p>
-              <h6 className="description card-text"><GiTabletopPlayers size={25} /> Players: {minPlayers} - {maxPlayers}</h6>
-              <p>{`Quantity: ${quantity}`}</p>
-              <button className="btn btn-outline-danger my-2 my-sm-0" onClick={handleRemoveProduct}>Remove from cart</button>
             </div>
           </div>
         </div>
