--- client/src/pages/Wishlist.css
@@ -1,3 +1,0 @@
-body {
-  background-image: url(../../public/clouds.jpeg);
-}
