+++ client/src/pages/Wishlist.jsx
@@ -1,4 +1,4 @@
-import './Wishlist.css';
+import './pagesCSS/Wishlist.css';
 export default function Wishlist() {
   return <div>Wishlist</div>
 }
