+++ client/src/pages/pagesCSS/Account.css
@@ -1,5 +1,5 @@
 body {
-  background-image: url(../../public/clouds.jpeg);
+  background-image: url(../../../public/clouds.jpeg);
 }
 
 .butcherman-link {
