+++ client/src/pages/pagesCSS/Cart.css
@@ -1,5 +1,5 @@
 body {
-  background-image: url(../../public/clouds.jpeg);
+  background-image: url(../../../public/clouds.jpeg);
 }
 
 .cartImage {
