+++ client/src/pages/pagesCSS/Catalog.css
@@ -1,5 +1,5 @@
 body {
-  background-image: url(../../public/clouds.jpeg);
+  background-image: url(../../../public/clouds.jpeg);
 }
 
 .catalog-img {
