+++ client/src/pages/pagesCSS/Home.css
@@ -1,5 +1,5 @@
 body {
-  background-image: url(../../public/clouds.jpeg);
+  background-image: url(../../../public/clouds.jpeg);
 }
 
 .title {
