+++ client/src/pages/pagesCSS/ProductDetails.css
@@ -1,5 +1,5 @@
 body {
-  background-image: url(../../public/clouds.jpeg);
+  background-image: url(../../../public/clouds.jpeg);
 }
 
 .product-detail-img {
