+++ client/src/pages/pagesCSS/SearchResult.css
