+++ client/src/pages/pagesCSS/Wishlist.css
@@ -1,0 +1,3 @@
+body {
+  background-image: url(../../../public/clouds.jpeg);
+}
